Add streaming send to csp.Chan and csp.SendCloser.
